+++ ThunderCats/TCInvalidArgument.m
@@ -1,8 +1,8 @@
 //
-//  UICollectionReusableView+TCAdditions.h
+//  TCInvalidArgument.h
 //  ThunderCats
 //
-//  Copyright (c) 2015 Metova Inc.
+//  Copyright (c) 2016 Metova Inc.
 //
 //  MIT License
 //
@@ -26,31 +26,13 @@
 //  WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
+#import "TCInvalidArgument.h"
 
-#import <UIKit/UIKit.h>
-#import "TCNullabilityMacros.h"
+@implementation TCInvalidArgument
 
-@interface UICollectionReusableView (TCAdditions)
-
-/**
- *  Returns a string equal to the class name for use as a reuse identifier.
- *
- *  @return A reuse identifier that is equal to the class name.
- */
-+ (NSString * __tc_null_unspecified)tc_reuseIdentifier;
-
-/**
- *  Returns a string equal to the class name.
- *
- *  @return A string equal to the class name.
- */
-+ (NSString * __tc_null_unspecified)tc_nibName;
-
-/**
- *  Returns a nib for the cell subclass if one exists and is named after the class.
- *
- *  @return The nib for the cell subclass.
- */
-+ (UINib * __tc_null_unspecified)tc_nib;
++ (void)raiseWithReason:(NSString *)reason
+{
+    [NSException raise:@"TCInvalidArgument" format:@"%@", reason];
+}
 
 @end