+++ ThunderCats/UIAlertController+TCAdditions.h
@@ -27,12 +27,60 @@
 //
 
 #import <UIKit/UIKit.h>
+#import "TCNullabilityMacros.h"
 
 @interface UIAlertController (TCAdditions)
 
 /**
+ *  A block definition for the cancelAction on the UIAlertController+TCAdditions methods.
+ */
+typedef void (^ TCCancelAction)(UIAlertAction * __tc_nonnull);
+
+/**
  *  This is a convenient method that returns an alert controller with the title, "Not Implemented", and message, "This feature is not yet implemented."
  */
-+ (instancetype)tc_notImplementedAlert;
++ (instancetype __tc_nonnull)tc_notImplementedAlert;
+
+/**
+ *  Creates a UIAlertController standard (popup) alert with a cancel button
+ *
+ *  @param title             The title of the alert
+ *  @param message           The message displayed on the alert
+ *  @param cancelButtonTitle The text displayed on the cancel button
+ *  @param cancelAction      A completion block which will be executed when the user taps cancel which has one UIAlertAction as a parameter. If the value is nil the alert will be dismissed upon tapping the cancel button.
+ *
+ *  @return A UIAlertController
+ */
++ (instancetype __tc_nonnull)tc_alertWithTitle:(NSString * __tc_nonnull)title
+                                       message:(NSString * __tc_nonnull)message
+                             cancelButtonTitle:(NSString * __tc_nonnull)cancelButtonTitle
+                                  cancelAction:(TCCancelAction __tc_nullable)cancelAction;
+
+
+/**
+ *  Creates a UIAlertController which uses an action sheet
+ *
+ *  @param title             The title of the alert
+ *  @param message           The message displayed on the alert
+ *  @param cancelButtonTitle The text displayed on the cancel button
+ *  @param cancelAction      A completion block which will be executed when the user taps cancel which has one UIAlertAction as a parameter. If the value is nil the alert will be dismissed upon tapping the cancel button.
+ *
+ *  @return A UIAlertController
+ */
++ (instancetype __tc_nonnull)tc_alertWithActionSheet:(NSString * __tc_nonnull)title
+                                             message:(NSString * __tc_nonnull)message
+                                   cancelButtonTitle:(NSString * __tc_nonnull)cancelButtonTitle
+                                        cancelAction:(TCCancelAction __tc_nullable)cancelAction;
+
+/**
+ *  Creates a UIAlertController which uses an action sheet for choosing from Camera Roll or Taking a Photo. Checks which source is available and only displays the available sources. Must instantiate and pass in an image picker with delegate already set.
+ *
+ *  @param imagePicker              The image picker to be used
+ *  @param presentingViewController The view controller presenting the image picker
+ *
+ *  @return A UIAlertController
+ */
++ (instancetype __tc_nonnull)tc_imagePickerActionSheet:(UIImagePickerController * __tc_nonnull)imagePicker
+                                    presentingViewController:(UIViewController * __tc_nonnull)presentingViewController;
 
 @end