+++ ThunderCats/UITextField+TCAdditions.h
@@ -34,8 +34,31 @@
 /**
  *  Sets the padding to the left and right sides of the text field.
  *
- *  @param paddingWidth The width to use for the padding.
+ *  @param paddingWidth The width to use for the padding, can not be negative
+ *  @exception TCInvalidArgument Throws TCInvalidArgument if a negative paddingWidth is passed
  */
 - (void)tc_setLeftAndRightPadding:(CGFloat)paddingWidth;
+
+/**
+ *  Adds a toolbar to keyboard with a next function which advances to the next textfield when tapped.
+ *
+ *  @param toolbarColor  The UIColor of the toolbar
+ *  @param nextTextField The textfield that should becomeFirstResponder next
+ */
+- (void)tc_addNextButtonToolbarWithColor:(UIColor * __tc_nonnull)toolbarColor
+                               nextField:(UITextField * __tc_nonnull)nextTextField;
+
+/**
+ *  Adds a toolbar to keyboard with a custom button that performs the action passed in
+ *
+ *  @param buttonTitle   The title of the button
+ *  @param toolbarColor  The UIColor of the toolbar
+ *  @param target The target of the action
+ *  @param selector      The action to be performed on button press
+ */
+- (void)tc_addToolbarWithButton:(NSString * __tc_nonnull)buttonTitle
+                   toolbarColor:(UIColor * __tc_nonnull)toolbarColor
+                         target:(id __tc_nullable)target
+                         action:(SEL __tc_nullable)selector;
 
 @end